refactor: Car 클래스를 main.js에서 car.js로 분리
